[1156] Updated types and renamed files
