+++ src/features/tasks/components/index.ts
@@ -1,5 +1,6 @@
 export * from "./CompletedTask";
 export * from "./DeleteAllScheduledTasks";
+export * from "./EditCompletedTaskForm";
 export * from "./EditInProgressTaskForm";
 export * from "./EditScheduledTaskForm";
 export * from "./ExampleCompletedTask";
